+++ mdast_util_to_markdown/src/construct_name.rs
@@ -1,5 +1,4 @@
 #[derive(Clone, PartialEq)]
-#[allow(dead_code)]
 pub enum ConstructName {
     Autolink,
     Blockquote,
