+++ mdast_util_to_markdown/src/handle/blockquote.rs
@@ -5,7 +5,6 @@ use crate::{
     construct_name::ConstructName,
     message::Message,
     state::{Info, State},
-    util::indent_lines::indent_lines,
 };
 
 use super::Handle;
@@ -19,20 +18,21 @@ impl Handle for Blockquote {
         node: &Node,
     ) -> Result<alloc::string::String, Message> {
         state.enter(ConstructName::Blockquote);
-        let value = indent_lines(&state.container_flow(node)?, map);
+        let value = state.container_flow(node)?;
+        let value = state.indent_lines(&value, map);
         Ok(value)
     }
 }
 
-fn map(value: &str, _line: usize, blank: bool) -> String {
+fn map(line: &str, _index: usize, blank: bool) -> String {
     let marker = ">";
-    let total_allocation = marker.len() + value.len() + 1;
+    let total_allocation = marker.len() + line.len() + 1;
     let mut result = String::with_capacity(total_allocation);
     result.push_str(marker);
     if !blank {
         let blank_str = " ";
         result.push_str(blank_str);
     }
-    result.push_str(value);
+    result.push_str(line);
     result
 }
