+++ mdast_util_to_markdown/src/util/mod.rs
@@ -10,7 +10,6 @@ pub mod check_strong;
 pub mod format_code_as_indented;
 pub mod format_heading_as_setext;
 pub mod format_link_as_auto_link;
-pub mod indent_lines;
 pub mod longest_char_streak;
 pub mod pattern_in_scope;
 pub mod safe;
